+++ tests/exp_voxel_convolution_test.py
@@ -15,7 +15,7 @@ def test_convolution(keys):
     @hk.without_apply_rng
     @hk.transform
     def c(x, z):
-        x = IrrepsArray.from_contiguous(irreps_in, x)
+        x = IrrepsArray.from_array(irreps_in, x)
         x = Convolution(
             irreps_out=irreps_out,
             irreps_sh=irreps_sh,
@@ -24,7 +24,7 @@ def test_convolution(keys):
             relative_starts={0: 0.0, 1: 0.0, 2: 0.5},
             steps=((1.0, 1.0, 1.0), z),
         )(x)
-        return x.contiguous
+        return x.array
 
     f = jax.jit(c.apply)
 
@@ -56,7 +56,7 @@ def test_convolution_defaults(keys):
     @hk.without_apply_rng
     @hk.transform
     def c(x):
-        x = IrrepsArray.from_contiguous(irreps_in, x)
+        x = IrrepsArray.from_array(irreps_in, x)
         x = Convolution(
             irreps_out=irreps_out,
             irreps_sh=irreps_sh,
@@ -64,7 +64,7 @@ def test_convolution_defaults(keys):
             num_radial_basis=3,
             steps=(1.0, 1.0, 1.0),
         )(x)
-        return x.contiguous
+        return x.array
 
     f = jax.jit(c.apply)